fix: 修复回调函数返回类型并添加nil检查
统一将回调函数的返回类型从int32改为uintptr以匹配Windows API要求
为所有回调函数调用添加nil检查，避免空指针崩溃
在generate_test.go中更新函数类型生成逻辑以支持新返回类型

+++ project/hyperdbg/sdk/HyperDbgSdk.go
@@ -179,9 +179,9 @@ type (
 	RegistersNames                                          = [120]*int8
 )
 
-type SendMessageWithParamCallback func(*int8) int32
+type SendMessageWithParamCallback func(*int8) uintptr
 
-type SendMessageWithSharedBufferCallback func() int32
+type SendMessageWithSharedBufferCallback func() uintptr
 
 type SymbolMapCallback func(uint64, *int8, *int8, uint32) uintptr
